Completely refactgored tests to async pattern
