--- tests/ElCamino.AspNet.Identity.AzureTable.Tests/ModelTests/IdentityRoleTests.cs
@@ -1,28 +1,0 @@
-﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
-
-using System;
-#if net45
-using ElCamino.AspNet.Identity.AzureTable.Model;
-#else
-using ElCamino.AspNetCore.Identity.AzureTable.Model;
-#endif
-using Xunit;
-
-namespace ElCamino.Web.Identity.AzureTable.Tests.ModelTests
-{
-    public class IdentityRoleTests
-    {
-        [Fact(DisplayName = "IdentityRoleSet_Id")]
-#if net45
-        [Trait("Identity.Azure.Model", "")]
-#else
-        [Trait("IdentityCore.Azure.Model", "")]
-#endif
-        public void IdentityRoleSet_Id()
-        {
-            var role = new IdentityRole();
-            role.Id = Guid.NewGuid().ToString();
-            Assert.Equal<string>(role.RowKey, role.Id);
-        }
-    }
-}
