--- tests/ElCamino.AspNet.Identity.AzureTable.Tests/ModelTests/IdentityUserClaimTests.cs
@@ -1,28 +1,0 @@
-﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
-
-using System;
-#if net45
-using ElCamino.AspNet.Identity.AzureTable.Model;
-#else
-using ElCamino.AspNetCore.Identity.AzureTable.Model;
-#endif
-using Xunit;
-
-namespace ElCamino.Web.Identity.AzureTable.Tests.ModelTests
-{
-    public class IdentityUserClaimTests
-    {
-        [Fact(DisplayName = "IdentityUserClaimGet_UserId")]
-#if net45
-        [Trait("Identity.Azure.Model", "")]
-#else
-        [Trait("IdentityCore.Azure.Model", "")]
-#endif
-        public void IdentityUserClaimGet_UserId()
-        {
-            var uc = new IdentityUserClaim();
-            uc.GenerateKeys();
-            Assert.Equal(uc.PartitionKey, uc.UserId);
-        }
-    }
-}
