+++ tests/ElCamino.AspNetCore.Identity.AzureTable.Tests/ModelTests/ApplicationUser.cs
@@ -1,0 +1,13 @@
+﻿// MIT License Copyright 2017 (c) David Melendez. All rights reserved. See License.txt in the project root for license information.
+
+using ElCamino.AspNetCore.Identity.AzureTable.Model;
+
+namespace ElCamino.Web.Identity.AzureTable.Tests.ModelTests
+{
+    public class ApplicationUser : IdentityUser
+    {
+        public string FirstName { get; set; }
+
+        public string LastName { get; set; }
+    }
+}
