+++ tests/ElCamino.AspNetCore.Identity.AzureTable.Tests/config.json
@@ -7,11 +7,11 @@
       "ConnectionString": "Server=(localdb)\\mssqllocaldb;Database=aspnet5-SampleMvc2-444042c2-53c3-4c2b-ae65-5c753b0d2e4a;Trusted_Connection=True;MultipleActiveResultSets=true"
     }
   },
-	"IdentityAzureTable": {
-		"identityConfiguration": {
-			"tablePrefix": "",
-			"storageConnectionString": "UseDevelopmentStorage=true;",
-			"locationMode": "PrimaryOnly"
-		}
-	}
+  "IdentityAzureTable": {
+    "identityConfiguration": {
+      "tablePrefix": "",
+      "storageConnectionString": "UseDevelopmentStorage=true;",
+      "locationMode": "PrimaryOnly"
+    }
+  }
 }
